feat(todo): rework animation set and stone model via Blockbench
- re-export todo_aoi.animation.json with 12 clips (6 new ability clips)
- boogie_woogie: anticipation at 0.0417s, palm contact at 0.0833s (2t)
- fake_clap: near-identical silhouette, recovery offset one grid slot
- new clips: stone_throw, momentum_strike, peak, revised (loop)
- new todo_stone pebble model (3 cubes) + geo.json + texture
- bbmodel sources updated under source-assets/blockbench/

+++ src/client/java/jujutsu/mod/client/character/todo/TodoStatusHud.java
@@ -99,7 +99,7 @@ public final class TodoStatusHud {
 		if (beat == 4) {
 			double pulse = 0.65 + 0.35 * (0.5 + 0.5
 					* Math.sin((level.getGameTime() + tickCounter.getGameTimeDeltaPartialTick(false)) * 0.22));
-			int peakAlpha = Math.round(255.0 * pulse);
+			int peakAlpha = Math.round((float) (255.0 * pulse));
 			peakColor = (peakAlpha << 24) | (PEAK_ACCENT & 0x00FFFFFF);
 			Component peak = Component.translatable("hud.jujutsumod.todo.rhythm.peak");
 			graphics.drawString(font, peak, x + width - font.width(peak) - 7, y + 5, peakColor, false);
